Final refactor of CPU kernels, some renaming/adjustments to GPU kernel args/params, naming improvements. Updated ReadME with latest benchmark results

+++ include/minfer/models/qwen3/module.hpp
@@ -21,7 +21,7 @@ public:
 
 private:
     void cpu_forward(float* x_out, uint32_t token_id);
-    void metal_forward(MTL::Buffer* x_out, int token_id);
+    void metal_forward(MTL::Buffer* x_out, uint32_t token_id);
 };
 
 class Qwen3LMHead : public Linear {
@@ -56,7 +56,7 @@ class Qwen3GQA : public GQA {
 public:
     Qwen3GQA(
         int block_idx, int d_model, int n_heads, int n_kv_heads, int d_head, int d_rotary,
-        size_t max_seq_len,
+        int max_seq_len,
         float eps, float freq_base,
         TPtr wq, TPtr wk, TPtr wv,
         TPtr wo, TPtr wq_norm, TPtr wk_norm,
